+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 * Usage of `lookup_existing` preventing propagation of resource `options` [[#372](https://github.com/okube-ai/laktory/issues/372)]
 * Typo in unity catalog quick start documentation 
 ### Updated
-* n/a
+* Databricks `group` resource properties
+* Databricks `user` resource properties
 ### Breaking changes
 * n/a